# docs/OpenApi.yaml
openapi: 3.0.3
info:
  title: DSS Building Maintenance API (FT UNNES)
  description: |
    API documentation for the Decision Support System (DSS) for Building Maintenance
    Prioritization at Fakultas Teknik, Universitas Negeri Semarang.

    **Base URL:** `http://localhost:3001`

    ## System Logic (SAW Method)

    The system uses the **Simple Additive Weighting (SAW)** method:

    1. **Criteria:**
       - `C1` — Building age (cost criterion — older = higher priority)
       - `C2` — Physical condition, composed of 3 sub-criteria:
         - `C21` — Structural condition (benefit: lower score = worse condition)
         - `C22` — Architectural condition (benefit)
         - `C23` — MEP (Mechanical, Electrical & Plumbing) condition (benefit)
       - `C3` — Utility / number of users (benefit)
       - `C4` — Damage impact level (benefit)
       - `C5` — Years since last maintenance (cost)

    2. **Normalization:** Min/Max normalization on a 0–1 scale.
       - **Benefit criterion:** `normalized = value / max_value`
       - **Cost criterion:** `normalized = min_value / value`

    3. **Weighting:** `FinalScore = Σ (normalized_criteria × weight)`

    4. **Sub-weight resolution:** Each sub-weight of C2 is multiplied by the parent
       weight before summing.

    ## Authentication

    All protected endpoints require a Bearer JWT token in the `Authorization` header:
    ```
    Authorization: Bearer <access_token>
    ```

  version: 1.3.0
  contact:
...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Access Token. Obtain from `POST /auth/login`.
  schemas:
...
    WebResponse:
      type: object
      required:
        - statusCode
        - success
        - message
      properties:
        statusCode:
          type: integer
          example: 200
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Success"
        data:
          nullable: true
          description: Response payload; null for delete/logout operations.
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Validation Error"
        error:
          type: string
          example: "Bad Request"

    # -------------------------
    # Auth
    # -------------------------
    RegistrationRequest:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          description: Alphanumeric username (used as display name)
          example: "admin_main"
        email:
          type: string
          format: email
          example: "admin@ft.unnes.ac.id"
        password:
          type: string
          description: Alphanumeric only (validated by class-validator @IsAlphanumeric)
          example: "StrongPass123"
        avatar:
          type: string
          nullable: true
          description: Optional avatar URL
          example: "https://example.com/avatar.png"
    RegisteredUser:
      type: object
      properties:
        id:
          type: string
          format: cuid
          example: "clxxxxxxxxxxxxxxxxxx"
        name:
          type: string
          example: "admin_main"
        email:
          type: string
          format: email
          example: "admin@ft.unnes.ac.id"
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "admin@ft.unnes.ac.id"
        password:
          type: string
          example: "StrongPass123"
    LoginUser:
      type: object
      properties:
        accessToken:
          type: string
          description: Short-lived JWT access token
          example: "eyJhbGci..."
        refreshToken:
          type: string
          description: Long-lived JWT refresh token (7-day expiry)
          example: "eyJhbGci..."
        username:
          type: string
          example: "admin_main"
        isAdmin:
          type: boolean
          example: false
        id:
          type: string
          format: cuid
          example: "clxxxxxxxxxxxxxxxxxx"
    CurrentUser:
      type: object
      properties:
        id:
          type: string
          format: cuid
          example: "clxxxxxxxxxxxxxxxxxx"
        name:
          type: string
          example: "admin_main"
        email:
          type: string
          format: email
          example: "admin@ft.unnes.ac.id"
        isAdmin:
          type: boolean
          example: false

    # -------------------------
    # Buildings
    # -------------------------
...
paths:
...
  /auth/register:
    post:
      summary: Register a new user
      description: |
        Creates a new user account. By default, new accounts are **not admin** (`isAdmin: false`).
        An admin account must be promoted manually in the database.
        Password must be **alphanumeric only** (letters and digits, no special characters).
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistrationRequest'
            example:
              username: "admin_main"
              email: "admin@ft.unnes.ac.id"
              password: "StrongPass123"
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/WebResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/RegisteredUser'
              example:
                statusCode: 201
                success: true
                message: "User registered successfully"
                data:
                  id: "clxxxxxxxxxxxxxxxxxx"
                  name: "admin_main"
                  email: "admin@ft.unnes.ac.id"
        '400':
          description: Validation error or email already in use
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                emailInUse:
                  summary: Email already registered
                  value:
                    statusCode: 400
                    success: false
                    message: "Email already in use"
                    error: "Bad Request"
                validationError:
                  summary: Invalid input fields
                  value:
                    statusCode: 400
                    success: false
                    message: ["password must contain only letters and numbers"]
                    error: "Bad Request"
  /auth/login:
    post:
      summary: Login and obtain tokens
      description: |
        Authenticates a user with email and password.
        Returns an **access token** (short-lived) and a **refresh token** (7-day expiry).
        The refresh token is stored server-side and must be presented to `/auth/refresh`.
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: "admin@ft.unnes.ac.id"
              password: "StrongPass123"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/WebResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LoginUser'
              example:
                statusCode: 200
                success: true
                message: "Success"
                data:
                  accessToken: "eyJhbGci..."
                  refreshToken: "eyJhbGci..."
                  username: "admin_main"
                  isAdmin: false
                  id: "clxxxxxxxxxxxxxxxxxx"
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                statusCode: 401
                success: false
                message: "Invalid credentials"
                error: "Unauthorized"
  /auth/refresh:
    post:
      summary: Refresh access token
      description: |
        Exchanges a valid **refresh token** for a new access token and refresh token pair.
        The refresh token must be sent in the `Authorization: Bearer <refresh_token>` header.
        Uses the `RefreshTokenGuard` (a separate JWT strategy that validates against the refresh secret).
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Tokens refreshed successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/WebResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LoginUser'
              example:
                statusCode: 200
                success: true
                message: "Success"
                data:
                  accessToken: "eyJhbGci_new..."
                  refreshToken: "eyJhbGci_new..."
                  username: "admin_main"
                  isAdmin: false
                  id: "clxxxxxxxxxxxxxxxxxx"
        '401':
          description: Refresh token invalid or expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/me:
    get:
      summary: Get current session user
      description: |
        Returns profile information of the currently authenticated user.
        Requires a valid **access token** in the Authorization header.
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user data
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/WebResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CurrentUser'
              example:
                statusCode: 200
                success: true
                message: "Success"
                data:
                  id: "clxxxxxxxxxxxxxxxxxx"
                  name: "admin_main"
                  email: "admin@ft.unnes.ac.id"
                  isAdmin: false
        '401':
          description: Access token invalid or missing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/session:
    delete:
      summary: Logout (invalidate session)
      description: |
        Invalidates the current session by clearing the stored refresh token from the database.
        Subsequent refresh requests will fail until the user logs in again.
        Requires a valid **access token**.
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logged out successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/WebResponse'
                  - type: object
                    properties:
                      data:
                        type: 'null'
              example:
                statusCode: 200
                success: true
                message: "Success"
                data: null
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ============================================================
  # BUILDINGS
  # ============================================================
...